rename react components to kebab-case
