+++ src/components/product-list/product-list.tsx
@@ -1,7 +1,7 @@
 'use client'
 import React, { useEffect, useState } from 'react'
 import { SuperMarkets } from '@/src/types/superMarkets'
-import style from '@/src/components/ProductList/ProductList.module.css'
+import style from '@/src/components/product-list/ProductList.module.css'
 import Product from '@/src/models/product'
 import ProductService from '@/src/services/product-service'
